remove opacity from hidden section
opacity and z-index values apply stacking contexts to the elements they are applied to. Stacking contexts are applied to a single root HTML element; child elements are confined to an order within this stacking context. Elements with the same stacking order are stacked according to appearance within the HTML. This was affecting the 'hidden' topic sections and the pop-up edit menus; they had similar stacking orders.